+++ src/ReadonlyRecord.ts
@@ -1,10 +1,11 @@
 /*
- * Various functions to aid in working with readonly `Record`s and more broadly
- * readonly objects.
+ * This module targets readonly objects in the sense of maps. For readonly
+ * objects in the sense of product types see the `ReadonlyStruct` module.
  *
  * @since 0.10.0
  */
-import { flow, pipe } from "fp-ts/function"
+
+import { flow } from "fp-ts/function"
 import { Predicate, not } from "fp-ts/Predicate"
 import { Endomorphism } from "fp-ts/Endomorphism"
 import { Option } from "fp-ts/Option"
@@ -47,107 +48,6 @@ export const lookupV =
   <A>(x: RR.ReadonlyRecord<string, A>) =>
   (k: string): Option<A> =>
     RR.lookup(k)(x)
-
-/**
- * Merge two records together. For merging many identical records, instead
- * consider defining a semigroup.
- *
- * @example
- * import { merge } from 'fp-ts-std/ReadonlyRecord';
- *
- * assert.deepStrictEqual(merge({ a: 1, b: 2 })({ b: 'two', c: true }), { a: 1, b: 'two', c: true });
- *
- * @since 0.10.0
- */
-export const merge =
-  <A>(x: A) =>
-  <B>(y: B): A & B => ({ ...x, ...y })
-
-/**
- * Pick a set of keys from a `Record`. The value-level equivalent of the `Pick`
- * type.
- *
- * @example
- * import { pick } from 'fp-ts-std/ReadonlyRecord';
- * import { pipe } from 'fp-ts/function'
- *
- * const picked = pipe(
- *   { a: 1, b: 'two', c: [true] },
- *   pick(['a', 'c'])
- * );
- *
- * assert.deepStrictEqual(picked, { a: 1, c: [true] });
- *
- * @since 0.10.0
- */
-export const pick =
-  <A, K extends keyof A>(ks: ReadonlyArray<K>) =>
-  (x: A): Pick<A, K> =>
-    // I don't believe there's any reasonable way to model this sort of
-    // transformation in the type system without an assertion - at least here
-    // it's in a single reused place
-    pipe(
-      ks,
-      RA.reduce({} as Pick<A, K>, (ys, k) =>
-        merge(ys)(k in x ? { [k]: x[k] } : {}),
-      ),
-    )
-
-//   {
-//   const o = {} as Pick<A, K>
-//
-//   /* eslint-disable */
-//   for (const k of ks) {
-//     o[k] = x[k]
-//   }
-//   /* eslint-enable */
-//
-//   return o
-// }
-
-/**
- * Like `pick`, but allows you to specify the input record upfront.
- *
- * @example
- * import { pickFrom } from 'fp-ts-std/ReadonlyRecord';
- *
- * type MyType = { a: number; b: string; c: ReadonlyArray<boolean> };
- * const picked = pickFrom<MyType>()(['a', 'c']);
- *
- * assert.deepStrictEqual(picked({ a: 1, b: 'two', c: [true] }), { a: 1, c: [true] });
- *
- * @since 0.12.0
- */
-export const pickFrom = <A>(): (<K extends keyof A>(
-  ks: ReadonlyArray<K>,
-) => (x: A) => Pick<A, K>) => pick
-
-/**
- * Omit a set of keys from a `Record`. The value-level equivalent of the `Omit`
- * type.
- *
- * @example
- * import { omit } from 'fp-ts-std/ReadonlyRecord';
- *
- * const sansB = omit(['b']);
- *
- * assert.deepStrictEqual(sansB({ a: 1, b: 'two', c: [true] }), { a: 1, c: [true] });
- *
- * @since 0.10.0
- */
-export const omit =
-  <K extends string>(ks: ReadonlyArray<K>) =>
-  <V, A extends RR.ReadonlyRecord<K, V>>(x: Partial<A>): Omit<A, K> => {
-    const y = { ...x }
-
-    /* eslint-disable */
-    for (const k of ks) {
-      delete y[k]
-    }
-    /* eslint-enable */
-
-    return y as Omit<A, K>
-  }
 
 /**
  * Filters out key/value pairs in the record for which the predicate upon the